rename flexknots.py to core.py
